refactor(python): split binding runtime and tests
